Refactor DOIManager to have authenticated client. Updated tests
