+++ src/DOIManager.php
@@ -15,6 +15,7 @@ class DOIManager
 {
 
     private $clientRepo = null;
+    private $authenticatedClient = null;
 
     /**
      * DOIManager constructor.
@@ -27,20 +28,57 @@ class DOIManager
 
     /**
      * Authenticate a client
+     * @param $appID
+     * @param null $sharedSecret
+     * @param null $ipAddress
      * @return bool
      */
-    public function authenticate()
-    {
-        return true;
+    public function authenticate(
+        $appID,
+        $sharedSecret = null,
+        $ipAddress = null
+    ) {
+        $client = $this->clientRepo->authenticate($appID, $sharedSecret, $ipAddress);
+
+        if ($client) {
+            $this->setAuthenticatedClient($client);
+            return true;
+        }
+
+        return false;
+
     }
 
     /**
-     * Returns the currently authenticated client
+     * @return null
      */
     public function getAuthenticatedClient()
     {
-
+        return $this->authenticatedClient;
     }
+
+    /**
+     * Setting the current authenticated client for this object
+     *
+     * @param $client
+     * @return $this
+     */
+    public function setAuthenticatedClient($client)
+    {
+        $this->authenticatedClient = $client;
+        return $this;
+    }
+
+    /**
+     * Returns if a client is authenticated
+     *
+     * @return bool
+     */
+    public function isClientAuthenticated()
+    {
+        return $this->getAuthenticatedClient() === null ? false : true;
+    }
+
 
     public function mint()
     {
