+++ tests/Repository/ClientRepositoryTest.php
@@ -1,5 +1,6 @@
 <?php
 
+use ANDS\DOI\Model\Client;
 use ANDS\DOI\Repository\ClientRespository;
 use Dotenv\Dotenv;
 
@@ -23,22 +24,13 @@ class ClientRepositoryTest extends PHPUnit_Framework_TestCase
     }
 
     /** @test */
-    public function it_should_show_a_client_is_authenticated_when_set()
-    {
-        $repo = $this->getClientRepository();
-        $repo->setAuthenticatedClient($repo->getFirst());
-        $this->assertNotNull($repo->isClientAuthenticated());
-    }
-
-    /** @test */
     public function it_should_authenticate_the_right_user_if_has_shared_secret()
     {
         $repo = $this->getClientRepository();
         $authenticate = $repo->authenticate(
             "94c5cdfc4183eca7f836f06f1ec5b85a4932758b", "04b51aa4aa"
         );
-        $this->assertTrue($authenticate);
-        $this->assertNotNull($repo->getAuthenticatedClient());
+        $this->assertNotFalse($authenticate);
     }
 
     /** @test */
@@ -49,18 +41,16 @@ class ClientRepositoryTest extends PHPUnit_Framework_TestCase
             "94c5cdfc4183eca7f836f06f1ec5b85a4932758b", "asdfasdfasdf"
         );
         $this->assertFalse($authenticate);
-        $this->assertNull($repo->getAuthenticatedClient());
     }
 
     /** @test */
     public function it_authenticates_user_if_ip_match_and_no_shared_secret_provided()
     {
         $repo = $this->getClientRepository();
-        $authenticate = $repo->authenticate(
+        $client = $repo->authenticate(
             "94c5cdfc4183eca7f836f06f1ec5b85a4932758b", null, "130.56.111.120"
         );
-        $this->assertTrue($authenticate);
-        $this->assertNotNull($repo->getAuthenticatedClient());
+        $this->assertInstanceOf(Client::class, $client);
     }
 
     /** @test */
@@ -71,7 +61,6 @@ class ClientRepositoryTest extends PHPUnit_Framework_TestCase
             "94c5cdfc4183eca7f836f06f1ec5b85a4932758b", null, "130.56.111.11"
         );
         $this->assertFalse($authenticate);
-        $this->assertNull($repo->getAuthenticatedClient());
     }
 
     /**
